refactor: make player a submodule
