+++ src/app/player/index.js
@@ -1,0 +1,3 @@
+export * from "./ai.js";
+export * from "./base.js";
+export * from "./human.js";
