+++ tests/index.html
@@ -19,7 +19,9 @@
     </script>
     <script type="module" src="./test_deck.js"></script>
     <script type="module" src="./test_game.js"></script>
-    <script type="module" src="./test_player.js"></script>
+    <script type="module" src="./player/test_base.js"></script>
+    <script type="module" src="./player/test_ai.js"></script>
+    <script type="module" src="./player/test_human.js"></script>
     <script type="module" src="./test_ui.js"></script>
     <script type="module" src="./ai/test_base.js"></script>
     <script type="module" src="./ai/test_combo.js"></script>
